Fixed a deadlock condition when adding many AVP devices
There exists a race condition when many different user space processes or
threads request new AVP devices from the vswitch and then begin to set the
link up on those devices while more AVP devices are being created.  The link
up request holds the rtnl_lock() mutex in the kernel and the new AVP creation
attempts to acquire the same rtnl_lock() from within register_netdev().  This
is not a problem in and of itself but because the link up request sends a
request to the vswitch and waits for a response a deadlock condition is
created as the vswitch can't respond until the ioctl() to add the AVP device
returns and that can't return because the rtnl_lock() is held by the link up.

In the interest of time (drop1) we have attempted to avoid this condition by
explicitly checking for the deadlock condition.  This needs to be revisited in
the future to separate the request and response threads in vswitch so that
these actions can be run in parallel.