Added support to AVP for per socket mbuf mempools
The AVP kernel device driver needs to be able to translate mbuf addresses from
multiple mempools (i.e., socket0, and socket1 in a NUMA enabled node).  To do
this it needs to be told how to translate incoming buffer addresses.  The AVP
API has been modified so that the host application specifies which pools to
use for socket0 and socket1.