Added jumbo frame support to AVP devices
AVP devices can now support chained mbufs.  The current maximum number of
mbufs in a chain is set to 5 to enable support for 9000 Bytes frames but can
be increased if 16K frames are needed in the future.  AVP devices are notified
about the largest possible MTU in the host and can adjust their own MTU
accordingly.