Updated AVP kernel driver to account for frame overhead
The "change MTU" netdevice callback was not accounting for the ethernet frame
overhead when determining whether an MTU change should be allowed.  The MTU
value passed in as an argument is always the payload length and must be
augmented by the header and CRC length when determining max frame size.