Updated WRS layer to facilitate exporting to customers
We need to be able to export our AVP kernel module and AVP PMD module as
standalone units for customers.  To do this effectively we need to minimize
dependencies that those module have on any parts of the existing DPDK.  This
update reorganizes some header files and updates some code so that it can be
compiled from within our DPDK tree and also as standalone units (e.g., the AVP
kernel module).