remove dual numa socket restriction
This update removes the restriction of only allowing 2 numa
nodes, and thus 2 mempools for the AVP devices.  The limit
is being changed to the maximum number of default DPDK
configuration of CONFIG_RTE_MAX_NUMA_NODES (8) numa nodes.

If the DPDK limit is increased, then the AVP mempool limit
should also be increased.