CGTS-706 - AVP kernel queue scheduler limit needs to be removed
The thread pool is defined as a per-cpu static structure,
therefore it can not be easily fragmented with individual
allocations. Therefore the limit (WRS_AVP_KTHREAD_MAX_RX_QUEUES)
is being increased significantly to avoid the possibility to
running out of resources.

In the future if this limit is hit again, it can be changed to
allocate threads from a pool and assign them on a per-cpu basis
to the cpu list.