Updating AVP device version to 1.0.0
The AVP device version number used to be a single value.  It was set to 6.  It
is now composed of release, major, and minor version numbers.  The new current
value is 1.0.0.  In the future, any minor updates to the device format that
does not require any driver change will be reflected in the minor version
number.  All other changes will be reflected in the release and/or major
version numbers and will require some form of backwards compatibility by the
driver.