AVP emulated vlan offload
Introduce a VLAN offload feature capability to pass the VLAN ID as
meta data between the guest and the host so that the guest and the
host do not need to manipulate the packet to insert and remove
VLAN headers.  This offers a significant performance improvement for
guest VLAN tagging since the guest vlan is immediately remapped to
a provider network VLAN segment, therefore the packet rewrite was
only required for the packet transit between the guest and the host.
With the offload feature, this unnecessary overhead is removed.