Fixes AVP kmod compile error in newer kernels
The following defines are no longer present in kernels newer than or equal to
3.10.x.  Looking at other DPDK kernel modules they have setup compatibility
macros to deal with compiling in older kernels.

  NETIF_F_HW_VLAN_TX
  NETIF_F_HW_VLAN_RX

In addition to the above defines the API signature of the following function
has changed in newer kernels.  The ethernet protocol is required as a 2nd
parameter in kernels newer than or equal to 3.10.x.

  __vlan_hwaccel_put_tag