Added CPU hotplug support to the AVP kernel driver
This commit updates the AVP kernel driver to add the ability to react to
logical onlining/offlining of a CPU.  When the AVP module is first loaded and
it discovers the AVP devices in the system, it assigns receives queues to each
online CPU.  Those queues are serviced by an AVP receive thread.  The receive
threads are designed under the assumption that there is only a single thread
per CPU.

When a CPU is offlined it is critical that the local receive thread is
terminated and the queues assigned to it be re-assigned to other receive
threads on other CPU instances rather than simply re-affining the receive
thread.  Simply re-affining the receive thread to another CPU would cause
memory access contention and possibly corruption as both threads would attempt
to access the same per-cpu structures.

This feature is dependent on the CONFIG_HOTPLUG_CPU configuration option.