Protected AVP thread array with mutex
The AVP per-cpu thread array and the thread objects were not protected by any
mutual exclusion mechanism.  Until now that has been sufficient but in
preparation for CPU hotplug we need to be able to support multiple processes
accessing the thread list simultaneously.

The avp_thread_lock mutex is used to protect any accesses to the avp_thread
list as well as creation/deletion of kthread objects.

The kthread->lock spinlock is still used to protected the per-thread queue
list and the rx_count value.