Split the AVP host and guest versions numbers
The version number associated to the host device format and guest driver
functionality were represented by the same #define in the AVP common header
file.  This is fine for the current development stream but when patching
features back to the field we need a way to control the guest driver version
number independently from the host device format number.  Since both the guest
and host software are built from the same repo we split the two version
numbers for clarity.  If a driver feature is patched back to the field then it
should not impact the host version number.