Updated AVP version numbers to v1.1.0 for vlan emulation
A previous commit introduced a guest driver feature that has a dependency on a
host device feature.  The guest driver should not be enabling this feature
unless the host device version is above the first version that supported this
functionality.  For now, this commit is simply adjusting the version number.
A future update will add the correct semantics in the driver.