CGTS-1738: avp: copy mac address before register netdev
AVP devices would be listed as having MAC 00:00:00:00:00:00 before being
adminstratively enabled.  This prevented being able to manipulate AVP devices
in UDEV rules based on MAC address.  This update changes the creation of AVP
devices so that the MAC address is copied to the net_device structure before
registering with the kernel.

This also limited the ability to create VLAN interfaces over AVP devices until
the link was administratively enabled and the MAC was available. This issue is
also now fixed.