CGTS-1738: avp: set parent device when created from pci device
AVP devices did not set their parent device pointer to indicate that they were
created as a result of a new PCI device being probed.  This had the
side-effect of limiting how AVP devices can be used in UDEV rules because it
was not possible to match on PCI address.  This commit sets the parent device
pointer to be the PCI device if the AVP device was created from a PCI device;
otherwise the parent device is not set (as it is today).

This is how the sysfs symlink looks for a PCI device.

    /sys/class/net/eth1 -> ../../devices/pci0000:00/0000:00:06.0/net/eth1

This is how the sysfs symlink looks for a non-PCI device.

    /sys/class/net/eth1 -> ../../devices/virtual/net/eth1