CGTS-1774: Export AVP host port speed via /sys/class/net/
The AVP devices had no support for ethtool operations, therefore
the speed could not be retrieved.  In order to enable the speed
a set of supported ethtool operations has been added to AVP
devices.  AVP devices can be queried for driver information,
settings, and channels.

In addition to the ethtool operations, the AVP devices have been
updated to properly report the operational/carrier state so that
is properly reflected in ip link and ethtool commands.