CGTS-2782: avp: only request vlan offload if supported by host
In a previous commit we updated the AVP kernel module to only leverage VLAN
hardware offload if the host AVP device supported this feature.  Unfortunately,
when this work was done it did not constrain when the offload flag was reported
to the host during configure operations.

The end result is that the driver reported the hardware offload flag in the
requested feature list even when the host did not support it (Release 1 host)
and the guest was not actually using it.  This caused the AVP configure request
to fail when an R2 guest was used on a R1 host.