CGTS-885: avp: potential deadlock in avp kernel module during avp create
The interactions between the AVS process and the AVP kernel module made it
possible for deadlock scenarios to completely lockup the node.  Specifically,
if AVS sends an ioctl down to the kernel at the same time as a userspace
process triggers a link up/down which generates a request up to AVS deadlock
may occur.  Originally, this happened because the AVS management thread is a
single task and it is unable to process any incoming notifications from the
kernel until the kernel completes any inprogress ioctl requests.  If these two
types of operations happen at exactly the right time deadlock can occur.

In release 1, an attempt was made to fix this case by adding a mutex mechanism
(avp_control_lock) between the ioctl() API and the AVS control request/response
queue such that an ioctl() could not begin if a userspace process already
initiated a notification up to AVS.  This eliminated the obvious deadlock on
the request queue but introduced a different, and more subtle, path to
deadlock.

When a userspace process makes a request to change the link state or MTU it
holds the rtnl_lock() before invoking the AVP netdev function pointer.  That
function then acquires the avp_control_lock (introduced above) before sending a
message up to AVS.  In the opposite direction if the AVS process initiates an
ioctl() call it will start by acquiring the avp_control_lock and then later the
rtnl_lock() when calling register_netdev().  Because the order of those lock
operations is different in the two scenarios deadlock can occur.  The fact that
rtnl_lock() was held by userspace before invoking the netdev callback functions
was unknown at the time of introducing the avp_control_lock and so this
possibility was overlooked.

scenario1:
    1) user space request (ip link set eth0 up)
    2) invokes rtnl_lock()
    3) invokes avp_net_open()
    4) invokes avp_ctrl_set_link_state()
    5) invokes down_read(avp_control_lock)
    6) sends message to AVS

scenario2:
    1) AVS sends ioctl to create new AVP device
    2) invokes avp_ioctl_create()
    3) invokes avp_dev_create()
    4) invokes avp_dev_configure()
    5) invoke rtnl_lock()

Since AVS has a single threaded management function there was no simple way to
address both scenarios.  Instead, the AVS to kernel interactions have been
changed so that the AVP host devices never send any requests up to AVS.  This
completely eliminates the possibility of deadlocks on the request queue or the
rtnl_lock() mutex.  The alternative is that AVS will poll the kernel for link
status changes for AVP host devices.  For all AVP guest devices the existing
asynchronous mechanism will continue to be used.