CGTS-3882: avp: packet validation for zero length or oversized packets
It is assumed that packets will not be zero length when received from the
application, but since this condition is possible and it causes AVS to segfault
we need to add additional validation.  Since neither a zero length packet or a
packet which is composed of more than 5 segments can be processed by AVS we are
filtering these out.

This issue is addressed in 4 places.

  1) In the QEMU backend AVP driver (in a separate commit)
  2) In the AVP kernel moduler driver (packet is dropped with log + stat)
  3) In the AVP PMD (packet is dropped with stat)
  4) In the AVS receive handler for AVP interfaces the "nb_segs" mbuf attribute
     is sanitized to always be at least 1.