CGTS-4245: avp: do not report valid checksums on received packets
The AVP kernel driver has been reporting CHECKSUM_UNNECESSARY on received
packets which tells the kernel that the packet checksums have already been
verified.  This was incorrect since AVS does not validate checksums before
passing packets up to the VM.  This setting is now CHECKSUM_NONE which tells
the kernel that checksums have not yet been validated and it is up to the
networking stack to check them.