CGTS-4659: avp: adjust rx queue mappings on migration attach
When a device is first configured a mapping is created so that each of the host
tx_queues are mapping to one of the rx_queues.  During a migration it is
possible that the host tx_queues change in number.  To ensure that the device
continues to service all queues the rx queue mappings need to be setup again.
In the AVP kernel module driver this is done automatically when the queues are
assigned to rx threads but the PMD did not handle this case until now.