wrs-avp: remove minimum ethernet length padding
The minimum Ethernet frame padding performed on packet transmit is
overwritting tracking data when performing benchmark tests since
the tests involve removing a vlan header that results in a packet
that is less than 60 bytes in the guest.  Since the guest does not
require padding out for a minimum frame size for virtualized NICs
the padding code is being removed to ensure the tracking data
remains unaffected.