net/avp: moving pmd under drivers directory
As part of the AVP upstreaming effort we need to move our AVP PMD in to the
proper DPDK drivers directory.  For the period of time that the driver needs to
live both in our repo and in the upstream repo (assuming it gets accepted) it
will be easier if they in the same file paths.  We may need to prop fixes to
and from the upstream repo to keep them aligned.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>