exec-env: bump guest driver version number
Bump the version number to v1.2.13 to align with the version that currently
exists in the titanium-cloud github repo.  The version number in that repo is
shared between the PMD and the Kernel module driver and so even though there
were no changes in the kernel drivers from version v1.2.8 onwards the version
was incremented due to changes in the PMD.  Going forward we can increment the
version separately but to avoid confusion with existing installations we'll
start off at the already released v1.2.13 version.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>