add license and readme files before release
The LICENSE and README files are migrated from the titanium-cloud github repo
from the 17.06 drivers directory.  Their contents have been updated to align
with the new compile and install methods for this repo.  Unlike the 17.06 build
instructions, the module must now be compiled against a compiled and installed
DPDK release in order to have access to the common AVP header files.