readme: updated build instructions with dpdk build
The kernel driver compilation is no longer standalone.  Now that the AVP PMD is
a part of the DPDK libraries it is no longer necessary to package the AVP
common header files (e.g., rte_avp_common.h) with the driver.  Instead, the
driver needs to be compiled against a set of installed DPDK library headers
which includes the AVP common header files.