use independent version number for kernel driver
Until now the DPDK PMD and the kernel driver have shared the same version
number, but since the two are now stored in separate repos we are going to
allow them to diverge when necessary.  The host version is still going to be
controlled by the DPDK shared library.

The last common point was 1.2.13 but now that we have introduced a fix to the
kernel driver we are diverging to 1.2.14.